Add ResourceView cache to Resource
