+++ Gems/Atom/RHI/Code/Source/RHI/FrameGraphCompileContext.cpp
@@ -5,12 +5,15 @@
  * SPDX-License-Identifier: Apache-2.0 OR MIT
  *
  */
-#include <Atom/RHI/FrameGraphCompileContext.h>
-#include <Atom/RHI/FrameGraphAttachmentDatabase.h>
 #include <Atom/RHI/Buffer.h>
 #include <Atom/RHI/BufferScopeAttachment.h>
+#include <Atom/RHI/BufferView.h>
+#include <Atom/RHI/FrameGraphAttachmentDatabase.h>
+#include <Atom/RHI/FrameGraphCompileContext.h>
 #include <Atom/RHI/Image.h>
 #include <Atom/RHI/ImageScopeAttachment.h>
+#include <Atom/RHI/ImageView.h>
+
 
 namespace AZ::RHI
 {
