+++ Gems/Atom/RHI/Code/Source/RHI/FrameGraphCompiler.cpp
@@ -6,20 +6,22 @@
  *
  */
 
-#include <Atom/RHI/FrameGraphCompiler.h>
 #include <Atom/RHI/BufferFrameAttachment.h>
 #include <Atom/RHI/BufferScopeAttachment.h>
+#include <Atom/RHI/BufferView.h>
 #include <Atom/RHI/Factory.h>
 #include <Atom/RHI/FrameGraph.h>
+#include <Atom/RHI/FrameGraphCompiler.h>
 #include <Atom/RHI/ImageFrameAttachment.h>
 #include <Atom/RHI/ImageScopeAttachment.h>
+#include <Atom/RHI/ImageView.h>
 #include <Atom/RHI/RHIUtils.h>
 #include <Atom/RHI/Scope.h>
 #include <Atom/RHI/SwapChainFrameAttachment.h>
 #include <Atom/RHI/TransientAttachmentPool.h>
 #include <AzCore/IO/SystemFile.h>
-#include <AzCore/std/sort.h>
 #include <AzCore/std/optional.h>
+#include <AzCore/std/sort.h>
 
 namespace AZ::RHI
 {
@@ -949,8 +951,20 @@ namespace AZ::RHI
                 {
                     const ImageViewDescriptor& imageViewDescriptor = node->GetDescriptor().m_imageViewDescriptor;
 
-                    // Multi device image views don't have a global cache, so we always cache them
-                    ImageView* imageView = GetImageViewFromLocalCache(image, imageViewDescriptor);
+                    ImageView* imageView = nullptr;
+                    // Check image's cache first as that contains views provided by higher level code.
+                    if (image->IsInResourceCache(imageViewDescriptor))
+                    {
+                        imageView = image->BuildImageView(imageViewDescriptor).get();
+                    }
+                    else
+                    {
+                        // If the higher level code has not provided a view, check local frame graph compiler's local cache.
+                        // The local cache is special and was mainly added to handle transient resources. This cache adds a dependency to
+                        // the resourceview ensuring they do not get deleted at the end of the frame and recreated at the start of the next
+                        // frame.
+                        imageView = GetImageViewFromLocalCache(image, imageViewDescriptor);
+                    }
 
                     node->SetImageView(imageView);
                 }
@@ -975,8 +989,20 @@ namespace AZ::RHI
                 {
                     const BufferViewDescriptor& bufferViewDescriptor = node->GetDescriptor().m_bufferViewDescriptor;
 
-                    // Multi device buffer views don't have a global cache, so we always cache them
-                    BufferView* bufferView = GetBufferViewFromLocalCache(buffer, bufferViewDescriptor);
+                    BufferView* bufferView = nullptr;
+                    // Check buffer's cache first as that contains views provided by higher level code.
+                    if (buffer->IsInResourceCache(bufferViewDescriptor))
+                    {
+                        bufferView = buffer->BuildBufferView(bufferViewDescriptor).get();
+                    }
+                    else
+                    {
+                        // If the higher level code has not provided a view, check local frame graph compiler's local cache.
+                        // The local cache is special and was mainly added to handle transient resources. This cache adds a dependency to
+                        // the resourceview ensuring they do not get deleted at the end of the frame and recreated at the start of the next
+                        // frame.
+                        bufferView = GetBufferViewFromLocalCache(buffer, bufferViewDescriptor);
+                    }
 
                     node->SetBufferView(bufferView);
                 }
