+++ Gems/Atom/RHI/Code/Source/RHI/Image.cpp
@@ -9,6 +9,7 @@
 #include <Atom/RHI/DeviceImageView.h>
 #include <Atom/RHI/Image.h>
 #include <Atom/RHI/ImageFrameAttachment.h>
+#include <Atom/RHI/ImageView.h>
 
 namespace AZ::RHI
 {
@@ -45,7 +46,7 @@ namespace AZ::RHI
 
     Ptr<ImageView> Image::BuildImageView(const ImageViewDescriptor& imageViewDescriptor)
     {
-        return aznew ImageView{ this, imageViewDescriptor, GetDeviceMask() };
+        return Base::GetResourceView(imageViewDescriptor);
     }
 
     uint32_t Image::GetResidentMipLevel() const
@@ -85,56 +86,6 @@ namespace AZ::RHI
     void Image::Shutdown()
     {
         Resource::Shutdown();
-    }
-
-    bool Image::IsInResourceCache(const ImageViewDescriptor& imageViewDescriptor)
-    {
-        bool isInResourceCache{true};
-        IterateObjects<DeviceImage>([&isInResourceCache, &imageViewDescriptor]([[maybe_unused]] auto deviceIndex, auto deviceImage)
-        {
-            isInResourceCache &= deviceImage->IsInResourceCache(imageViewDescriptor);
-        });
-        return isInResourceCache;
-    }
-
-    //! Given a device index, return the corresponding DeviceBufferView for the selected device
-    const RHI::Ptr<RHI::DeviceImageView> ImageView::GetDeviceImageView(int deviceIndex) const
-    {
-        AZStd::lock_guard lock(m_imageViewMutex);
-
-        if (m_image->GetDeviceMask() != m_deviceMask)
-        {
-            m_deviceMask = m_image->GetDeviceMask();
-
-            MultiDeviceObject::IterateDevices(
-                m_deviceMask,
-                [this](int deviceIndex)
-                {
-                    if (auto it{ m_cache.find(deviceIndex) }; it != m_cache.end())
-                    {
-                        m_cache.erase(it);
-                    }
-                    return true;
-                });
-        }
-
-        auto iterator{ m_cache.find(deviceIndex) };
-        if (iterator == m_cache.end())
-        {
-            //! Image view is not yet in the cache
-            auto [new_iterator, inserted]{ m_cache.insert(
-                AZStd::make_pair(deviceIndex, m_image->GetDeviceImage(deviceIndex)->GetImageView(m_descriptor))) };
-            if (inserted)
-            {
-                return new_iterator->second;
-            }
-        }
-        else if (&iterator->second->GetImage() != m_image->GetDeviceImage(deviceIndex).get())
-        {
-            iterator->second = m_image->GetDeviceImage(deviceIndex)->GetImageView(m_descriptor);
-        }
-
-        return iterator->second;
     }
 
     void ImageSubresourceLayout::Init(MultiDevice::DeviceMask deviceMask, const DeviceImageSubresourceLayout &deviceLayout)
