+++ Gems/Atom/RHI/Code/Source/RHI/Resource.cpp
@@ -7,12 +7,12 @@
  */
 #include <Atom/RHI.Reflect/BufferViewDescriptor.h>
 #include <Atom/RHI.Reflect/ImageViewDescriptor.h>
-#include <Atom/RHI/DeviceBufferView.h>
+#include <Atom/RHI/BufferView.h>
+#include <Atom/RHI/DeviceResourceView.h>
 #include <Atom/RHI/Factory.h>
-#include <Atom/RHI/DeviceImageView.h>
+#include <Atom/RHI/ImageView.h>
 #include <Atom/RHI/Resource.h>
 #include <Atom/RHI/ResourcePool.h>
-#include <Atom/RHI/DeviceResourceView.h>
 #include <AzCore/std/containers/unordered_map.h>
 #include <AzCore/std/hash.h>
 
@@ -119,5 +119,154 @@ namespace AZ::RHI
             m_pool->ShutdownResource(this);
         }
         MultiDeviceObject::Shutdown();
+    }
+
+    Ptr<ImageView> Resource::GetResourceView(const ImageViewDescriptor& imageViewDescriptor) const
+    {
+        const HashValue64 hash = imageViewDescriptor.GetHash();
+        AZStd::lock_guard<AZStd::mutex> registryLock(m_cacheMutex);
+        auto it = m_resourceViewCache.find(static_cast<uint64_t>(hash));
+        if (it == m_resourceViewCache.end())
+        {
+            return InsertNewImageView(hash, imageViewDescriptor);
+        }
+        else
+        {
+            // We've found a matching ResourceView in the cache, but another thread may be releasing the last intrusive_ptr while
+            // we are in this function, dropping the refcount to 0 (and forcing it to -1 for good measure) then deleting it.
+            //
+            //  There are 2 scenarios:
+            //
+            // m_useCount is -1. The other thread is already on the path to deleting it. We need to make a new one here
+            // and replace the old one.
+            //
+            // m_useCount is >=0. We cannot guarantee another thread won't drop the refcount to 0 after we check the value here
+            // so before we create a new intrusive_ptr, we need to use fetch_add to increment the refcount as we check the value to
+            // prevent a race
+            int useCount = it->second->m_useCount.fetch_add(2);
+            if (useCount == -1)
+            {
+                // The useCount was -1 before we incremented.
+                // Another thread is going to come along and delete the one we just found.
+                // Go ahead and erase it and insert a new one instead
+                m_resourceViewCache.erase(it);
+
+                return InsertNewImageView(hash, imageViewDescriptor);
+            }
+            else
+            {
+                // Create the new Ptr, increasing the refcount
+                Ptr<ImageView> result = static_cast<ImageView*>(it->second);
+
+                // Before we checked the value we artificially incremented the refcount to prevent another
+                // thread from letting it go to 0 again. Get rid of that artificial increase now that we have
+                // our new Ptr to hold on to the refcount
+                it->second->m_useCount.fetch_sub(2);
+                return result;
+            }
+        }
+    }
+
+    Ptr<BufferView> Resource::GetResourceView(const BufferViewDescriptor& bufferViewDescriptor) const
+    {
+        const HashValue64 hash = bufferViewDescriptor.GetHash();
+        AZStd::lock_guard<AZStd::mutex> registryLock(m_cacheMutex);
+        auto it = m_resourceViewCache.find(static_cast<uint64_t>(hash));
+        if (it == m_resourceViewCache.end())
+        {
+            return InsertNewBufferView(hash, bufferViewDescriptor);
+        }
+        else
+        {
+            // We've found a matching ResourceView in the cache, but another thread may be releasing the last intrusive_ptr while
+            // we are in this function, dropping the refcount to 0 (and forcing it to -1 for good measure) then deleting it.
+            //
+            //  There are 2 scenarios:
+            //
+            // m_useCount is -1. The other thread is already on the path to deleting it. We need to make a new one here
+            // and replace the old one.
+            //
+            // m_useCount is >=0. We cannot guarantee another thread won't drop the refcount to 0 after we check the value here
+            // so before we create a new intrusive_ptr, we need to use fetch_add to increment the refcount as we check the value to prevent
+            // a race
+            int useCount = it->second->m_useCount.fetch_add(2);
+            if (useCount == -1)
+            {
+                // The useCount was -1 before we incremented.
+                // Another thread is going to come along and delete the one we just found.
+                // Go ahead and erase it and insert a new one instead
+                m_resourceViewCache.erase(it);
+
+                return InsertNewBufferView(hash, bufferViewDescriptor);
+            }
+            else
+            {
+                // Create the new Ptr, increasing the refcount
+                Ptr<BufferView> result = static_cast<BufferView*>(it->second);
+
+                // Before we checked the value we artificially incremented the refcount to prevent another
+                // thread from letting it go to 0 again. Get rid of that artificial increase now that we have
+                // our new Ptr to hold on to the refcount
+                it->second->m_useCount.fetch_sub(2);
+                return result;
+            }
+        }
+    }
+
+    Ptr<ImageView> Resource::InsertNewImageView(HashValue64 hash, const ImageViewDescriptor& imageViewDescriptor) const
+    {
+        Ptr<ImageView> imageViewPtr = aznew ImageView{ static_cast<const RHI::Image*>(this), imageViewDescriptor, GetDeviceMask() };
+        if (imageViewPtr)
+        {
+            m_resourceViewCache[static_cast<uint64_t>(hash)] = static_cast<ResourceView*>(imageViewPtr.get());
+            return imageViewPtr;
+        }
+        else
+        {
+            return nullptr;
+        }
+    }
+
+    Ptr<BufferView> Resource::InsertNewBufferView(HashValue64 hash, const BufferViewDescriptor& bufferViewDescriptor) const
+    {
+        Ptr<BufferView> bufferViewPtr = aznew BufferView{static_cast<const RHI::Buffer*>(this), bufferViewDescriptor, GetDeviceMask()};
+        if (bufferViewPtr)
+        {
+            m_resourceViewCache[static_cast<uint64_t>(hash)] = static_cast<ResourceView*>(bufferViewPtr.get());
+            return bufferViewPtr;
+        }
+        else
+        {
+            return nullptr;
+        }
+    }
+
+    void Resource::EraseResourceView(ResourceView* resourceView) const
+    {
+        AZStd::lock_guard<AZStd::mutex> registryLock(m_cacheMutex);
+        auto itr = m_resourceViewCache.begin();
+        while (itr != m_resourceViewCache.end())
+        {
+            if (itr->second == resourceView)
+            {
+                m_resourceViewCache.erase(itr->first);
+                break;
+            }
+            itr++;
+        }
+    }
+
+    bool Resource::IsInResourceCache(const ImageViewDescriptor& imageViewDescriptor)
+    {
+        const HashValue64 hash = imageViewDescriptor.GetHash();
+        auto it = m_resourceViewCache.find(static_cast<uint64_t>(hash));
+        return it != m_resourceViewCache.end();
+    }
+
+    bool Resource::IsInResourceCache(const BufferViewDescriptor& bufferViewDescriptor)
+    {
+        const HashValue64 hash = bufferViewDescriptor.GetHash();
+        auto it = m_resourceViewCache.find(static_cast<uint64_t>(hash));
+        return it != m_resourceViewCache.end();
     }
 } // namespace AZ::RHI
