+++ Gems/Atom/RPI/Code/Include/Atom/RPI.Reflect/Image/Image.h
@@ -10,9 +10,11 @@
 
 #include <AzCore/Memory/SystemAllocator.h>
 
+#include <Atom/RHI/DeviceImageView.h>
 #include <Atom/RHI/Image.h>
 #include <Atom/RHI/ImagePool.h>
-#include <Atom/RHI/DeviceImageView.h>
+#include <Atom/RHI/ImageView.h>
+
 
 #include <Atom/RPI.Reflect/Configuration.h>
